Improved the update system to support three different modes (#152)
- Silent (auto-update in background without any prompts)
- Notify (if an update exists, show a notification message)
- Never (never search for an update)

The default update mode is now notify, because if it would be silent an older version would instantly update itself to a newer one. This would break the possibility for the user to stay on that older version without disconnecting from the world wide web.
The default update check interval was set to 24h from 12h - that's often enough.
Some typos fixed.