Installation in Portuguese
Olá,

Sorry for being such a newbie.

This is my first pull ever. I think I got the basics of how to make it.

I added the lines and translation to the **setup.iss** and the **products.iss** to portuguese.

I also added some information to the **readme.me** that I think it's missing.