Rewrite completely the listing of devices
Rewrite the caching of the information. Stop using the MMDevice, only use it to gather the needed date.

MMDevice is great to list devices, not to be used everywhere as I was doing since it's a COM Object.

I created a DeviceFullInfo that contains exactly what the app needs to work. This way it's lighter and resolve the lag issues.

Fixes #298
Fixes #293